+++ src/main/java/com/taishan/netty/vo/Header.java
@@ -7,31 +7,31 @@ public class Header {
     /**
      * 数据包长度
      */
-    private short length;
+    private int length;
 
     /**
      * 协议版本号
      */
-    private byte version = 1;
+    private short version = 1;
 
     /**
      * 序列号
      */
-    private short serialNumber;
+    private int serialNumber;
 
     /**
      * 接收方地址
      */
-    private short receiveAddr;
+    private int receiveAddr;
 
     /**
      * 发送方地址
      */
-    private short sendAddr;
+    private int sendAddr;
 
     /**
      * 命令码
      */
-    private byte opCode;
+    private short opCode;
 
 }
